requirements.txt
# 数据获取相关
tushare>=1.2.89
yfinance>=0.2.18
akshare>=1.12.0
requests>=2.31.0

# 数据处理和分析
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0

# 数据库相关
PyMySQL>=1.1.0
sqlalchemy>=2.0.0

# 配置和日志
PyYAML>=6.0
coloredlogs>=15.0
loguru>=0.7.0

# 图表和可视化
matplotlib>=3.7.0
plotly>=5.15.0
seaborn>=0.12.0

# 时间处理
python-dateutil>=2.8.2

# 技术分析
ta-lib>=0.4.25

# 数据验证
pydantic>=2.0.0

# 开发和测试工具
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0

# 其他工具
tqdm>=4.65.0
click>=8.1.0